Modify data types for FSkillLevelUpInfo

+++ Source/EOD/Statics/CharacterLibrary.h
@@ -243,11 +243,11 @@ public:
 
 	/** Minimum stamina required to use this skill */
 	UPROPERTY(EditDefaultsOnly)
-	int StaminaRequired;
+	int32 StaminaRequired;
 	
 	/** Minimum mana required to use this skill */
 	UPROPERTY(EditDefaultsOnly)
-	int ManaRequired;
+	int32 ManaRequired;
 	
 	/** Skill cooldown */
 	UPROPERTY(EditDefaultsOnly)
@@ -255,7 +255,7 @@ public:
 
 	/** Damage in percentage of player's magickal or physical attack that will inflicted on enemy */
 	UPROPERTY(EditDefaultsOnly)
-	float DamagePercent;
+	int32 DamagePercent;
 
 	/** Determines if this skill can be blocked */
 	UPROPERTY(EditDefaultsOnly)
